jinahub/rankers/LightGBMRanker/tests/requirements.txt
jina-commons @ git+https://github.com/jina-ai/jina-commons.git#egg=jina-commons
pytest==6.2.4